requirements.txt
absl-py==1.0.0
astor==0.8.1
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bce-python-sdk==0.8.64
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.9
click==8.0.3
cssselect==1.1.0
cssutils==2.3.0
cycler==0.11.0
Cython==0.29.26
dataclasses==0.8
decorator==4.4.2
distlib==0.3.4
entrypoints==0.3
et-xmlfile==1.1.0
fasttext==0.9.1
filelock==3.4.0
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
identify==2.4.0
idna==3.3
imageio==2.13.3
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.2.3
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.1
lmdb==1.2.1
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
nest-asyncio==1.5.4
networkx==2.5.1
nodeenv==1.6.0
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.4.0.46
opencv-python==4.5.4.60
openpyxl==3.0.9
paddleocr==2.3.0.2
paddlepaddle-gpu==2.2.1
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.16.0
premailer==3.10.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.12.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.0
shellcheck-py==0.8.0.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
tifffile==2020.9.3
toml==0.10.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
visualdl==2.2.2
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
